Refactor the build script
