+++ package.json
@@ -4,13 +4,11 @@
     "build": "vite build",
     "build:use-remote-models": "vite build -- use-remote-models",
     "dev": "vite --port 9999 --host -- use-remote-models",
-    "copy-models": "node copy_models.js",
     "css-update": "npx tailwindcss -i ./style/tailwind-input.css -o ./style/tailwind.css --watch",
     "generate-ssl": "openssl req -newkey rsa:2048 -new -nodes -x509 -days 3650 -keyout key.pem -out cert.pem",
-    "fetch-models": "node fetch_models.js",
-    "prod:github": "npm run build:use-remote-models && node build_sub_proj.js github && npm run copy-models",
-    "prod:use-remote-models": "npm run build:use-remote-models && node build_sub_proj.js && npm run copy-models",
-    "prod": "npm run build && npm run fetch-models && node build_sub_proj.js && npm run copy-models",
+    "prod:github": "npm run build:use-remote-models && node build_scripts.js build-sub-projects --github && node build_scripts.js copy-models",
+    "prod:use-remote-models": "npm run build:use-remote-models && node build_scripts.js build-sub-projects && node build_scripts.js copy-models",
+    "prod": "npm run build && node build_scripts.js fetch-models && node build_scripts.js build-sub-projects && node build_scripts.js copy-models",
     "startup": "cd ./dist && http-server -S -C ../cert.pem -K ../key.pem -o",
     "lint": "eslint .",
     "lint-fix": "eslint . --fix"
